requirements.txt
pandas==2.2.1
openpyxl==3.1.2
duckdb==1.0.0
stream-unzip==0.0.91
loguru==0.7.2
psutil==5.9.8
pyarrow==15.0.2
boto3==1.34.77
memory-profiler==0.61.0
pydantic==2.7.0
requests==2.31.0
dbt-duckdb==1.8.0
python-dotenv==1.0.1
tqdm==4.66.4
numpy==1.26.4
GDAL==3.4.1
shapely==2.0.4
geopandas==0.14.4
fiona==1.9.6
beautifulsoup4==4.12.3
xlrd==2.0.1
msoffcrypto-tool==5.4.1
